+++ jk_epg/custom_components/jk_epg/const.py
@@ -1,0 +1,5 @@
+"""Constants for JK EPG."""
+
+DOMAIN = "jk_epg"
+PLATFORMS = ["sensor", "calendar"]
+DEFAULT_URL = "http://1815012f-jk-epg:8099"